# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Mini Blog",
        "description": "Documentation Swagger de l'API Mini Blog",
        "contact": {
            "email": "admin@miniblog.test"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Liste des articles",
                "description": "Récupère la liste paginée des articles",
                "operationId": "93ded9e31119e6bbb22455c72ef97f19",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "titre": {
                                                        "type": "string"
                                                    },
                                                    "slug": {
                                                        "type": "string"
                                                    },
                                                    "contenu": {
                                                        "type": "string"
                                                    },
                                                    "image_url": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Créer un nouvel article",
                "operationId": "75bdc5f953261e6f34852386c9b522c0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "titre",
                                    "contenu"
                                ],
                                "properties": {
                                    "titre": {
                                        "type": "string",
                                        "example": "Mon premier article"
                                    },
                                    "contenu": {
                                        "type": "string",
                                        "example": "Contenu de l'article..."
                                    },
                                    "image": {
                                        "description": "Image de l'article",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article créé avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Afficher un article",
                "operationId": "a85697b2464e9aaa3f9980a83edbef88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'article"
                    },
                    "404": {
                        "description": "Article introuvable"
                    }
                }
            },
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Modifier un article",
                "operationId": "beb39fc2f8d02c16b41438aadbd9ba03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "titre": {
                                        "type": "string"
                                    },
                                    "contenu": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article modifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Supprimer un article",
                "operationId": "804276a4147e577744e158b2ce2b58f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion utilisateur",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Utilisateur connecté",
                "operationId": "8367feff40ba10e43187792cad3132c1",
                "responses": {
                    "200": {
                        "description": "Données utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Déconnexion réussie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/articles/{id}/commentaires": {
            "get": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Lister les commentaires d’un article",
                "operationId": "d635a4717dd573e0acc24a52176ef1af",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des commentaires",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "contenu": {
                                                "type": "string"
                                            },
                                            "valide": {
                                                "type": "boolean"
                                            },
                                            "article_id": {
                                                "type": "integer"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/commentaires": {
            "post": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Créer un commentaire",
                "operationId": "e0dea7de16a343ed452dd63a2b6aeb6f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "contenu",
                                    "article_id"
                                ],
                                "properties": {
                                    "contenu": {
                                        "type": "string",
                                        "example": "Très bon article !"
                                    },
                                    "article_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commentaire créé"
                    }
                }
            }
        },
        "/api/commentaires/{id}/valider": {
            "put": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Valider un commentaire",
                "operationId": "67b84410e63e55cbd9bf56c53be879a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du commentaire",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commentaire validé"
                    },
                    "404": {
                        "description": "Commentaire introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/commentaires/{id}": {
            "delete": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Supprimer un commentaire",
                "operationId": "5e72d322cf4c284ea9519a5b3cd5d832",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du commentaire",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commentaire supprimé"
                    },
                    "404": {
                        "description": "Commentaire introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "title": "Utilisateur",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "401": {
                "description": "Non authentifié. Veuillez fournir un token valide."
            },
            "403": {
                "description": "Accès refusé. L'utilisateur n’a pas les droits nécessaires (ex : admin)."
            }
        }
    },
    "tags": [
        {
            "name": "Articles",
            "description": "Articles"
        },
        {
            "name": "Authentification",
            "description": "Authentification"
        },
        {
            "name": "Commentaires",
            "description": "Commentaires"
        }
    ]
}